+++ crates/pytauri-core/CHANGELOG.md
@@ -4,7 +4,8 @@
 
 ### Added
 
-- [#46](https://github.com/WSH032/pytauri/pull/46) - add `webview::WebviewWindow`, `Manager`, `ImplManager`, `App::handle`
+- [#46](https://github.com/WSH032/pytauri/pull/46) - feat: add `webview::WebviewWindow`, `Manager`, `ImplManager`, `App::handle`.
+- [#48](https://github.com/WSH032/pytauri/pull/48) - feat: accessing the `WebviewWindow` in `Commands`.
 
 ## [0.1.0-beta.1]